cherrypy starter kit with postgres/backend/requirements.txt
CherryPy-SQLAlchemy
CherryPy
psycopg2-binary
marshmallow
python-dotenv
cherrys
hiredis
alembic
celery
flower

